+++ .eslintrc.js
@@ -17,6 +17,8 @@ module.exports = {
       join(__dirname, 'packages/dataset-sdk/tsconfig.json'),
       join(__dirname, 'packages/quickjs/tsconfig.json'),
       join(__dirname, 'tests/tsconfig.json'),
+      join(__dirname, 'framework/core/tsconfig.json'),
+      join(__dirname, 'framework/runtime/tsconfig.json'),
     ],
   },
   rules: {