+++ framework/runtime/src/types/agent.ts
@@ -1,0 +1,3 @@
+import { Runnable } from './runnable';
+
+export interface Agent<I extends { [key: string]: any }, O> extends Runnable<I, O> {}