+++ packages/ai-runtime/src/core/executor/base.ts
@@ -650,7 +650,26 @@ export abstract class AgentExecutorBase<T> {
           llmInputMessages: Joi.array().items(
             Joi.object({
               role: Joi.string().valid('system', 'user', 'assistant').empty([null, '']).default('user'),
-              content: Joi.string().required(),
+              content: Joi.alternatives(
+                Joi.string().allow(null, ''),
+                Joi.array().items(
+                  Joi.object({
+                    type: Joi.string().valid('text', 'image_url').required(),
+                  })
+                    .when(Joi.object({ type: Joi.valid('text') }).unknown(), {
+                      then: Joi.object({
+                        text: Joi.string().required(),
+                      }),
+                    })
+                    .when(Joi.object({ type: Joi.valid('image_url') }).unknown(), {
+                      then: Joi.object({
+                        imageUrl: Joi.object({
+                          url: Joi.string().required(),
+                        }).required(),
+                      }),
+                    })
+                )
+              ).required(),
               name: Joi.string().empty([null, '']),
             })
           ),